# backend-MyKP/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyKP API Documentation",
        "description": "API documentation for the MyKP backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "List notifications and mark them all as read. Two kinds are interleaved: 'notification' (new activity created) and 'reminder' (3/2/1 days before a bookmarked activity's registration deadline).",
                "operationId": "911565ad9aff9ba7eded3f189509e578",
                "responses": {
                    "200": {
                        "description": "OK — sorted newest first. Calling this endpoint marks every returned item as read.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Stable ID. 'act-{ActivityID}' for new-activity notifications, 'bm-{BookmarkID}-{days}d' for deadline reminders.",
                                                "type": "string",
                                                "example": "act-1"
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "reminder",
                                                "enum": [
                                                    "notification",
                                                    "reminder"
                                                ]
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Registration deadline in 3 days"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Seminar Bela Negara & Anti Narkoba 2026"
                                            },
                                            "fired_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-05-15T00:00:00+07:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/unread-count": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Number of notifications the user hasn't viewed yet (drives the bell badge on the dashboard)",
                "operationId": "0272623bed30dd4d72141741fb3f25da",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "description": "Number of currently-visible notifications not yet acknowledged by reading /api/notifications",
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum API Token",
                "bearerFormat": "string",
                "scheme": "bearer"
            }
        }
    },
...
}